# libs/contracts/openapi/gateway.yaml
openapi: 3.0.3
info:
  title: DocVault API Gateway
  description: |
    Unified API gateway for DocVault — secure document management system.

    ## Authentication

    **Development (localhost):** After signing in via `/auth/login`, the gateway sets an
    `dv_access_token` cookie. Subsequent API calls automatically include this cookie.

    **Production:** Pass the Keycloak JWT in the `Authorization` header:
    `Authorization: Bearer <token>`

    Both methods are accepted on all protected endpoints.

    ## Roles
    | Role | Permissions |
    |------|-------------|
    | `viewer` | Read published documents, download files |
    | `editor` | Create, upload, submit, archive documents |
    | `approver` | Approve/reject pending documents |
    | `compliance_officer` | Read audit logs and allowed metadata; cannot preview, stream, presign, or download file content |
    | `admin` | Full access |

    ## Document Lifecycle
    ```
    DRAFT → PENDING → PUBLISHED → ARCHIVED
    ```

    ## Compliance Officer Restriction
    Compliance officers (`compliance_officer`) are **never allowed** to preview,
    stream, presign, or download file content, regardless of ACL configuration —
    enforced by route guards and policy.service.ts.
  version: 0.1.0
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.docvault.internal
    description: Production
security:
  - bearerAuth: []
  - cookieAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Keycloak-issued RS256 JWT token passed as `Authorization: Bearer <token>`"
    cookieAuth:
      type: apiKey
      in: cookie
      name: dv_access_token
      description: JWT issued by the gateway after Keycloak OIDC login — set automatically by the frontend
  schemas:
...
    ClassificationLevel:
      type: string
      enum: [PUBLIC, INTERNAL, CONFIDENTIAL, SECRET]
      description: |
        Document security classification level.
        - PUBLIC: Anyone authenticated can view and download
        - INTERNAL: viewer+ role, must be PUBLISHED
        - CONFIDENTIAL: editor+ role, PUBLISHED, owner or explicit ACL required
        - SECRET: approver+ role, PUBLISHED, owner or explicit ACL required
        CONFIDENTIAL and SECRET documents are watermarked on download.

    # ─── Documents ─────────────────────────────────────────────────────────────
...
    AuditEventResult:
      type: string
      enum: [SUCCESS, DENY, CONFLICT, ERROR]
    AuditEvent:
      type: object
      properties:
        eventId: { type: string }
        timestamp: { type: string, format: date-time }
        actorId: { type: string }
        actorRoles: { type: array, items: { type: string } }
        action:
          type: string
          description: |
            Canonical audit action constants:
            COMMENT_ADDED, COMMENT_UPDATED, COMMENT_DELETED,
            COMMENT_UPDATE_DENIED, COMMENT_DELETE_DENIED,
            DOCUMENT_CREATED, DOCUMENT_METADATA_UPDATED,
            DOCUMENT_ACL_UPDATED, DOCUMENT_ACL_DELETED,
            DOCUMENT_SUBMIT, DOCUMENT_APPROVE, DOCUMENT_REJECT, DOCUMENT_ARCHIVE,
            DOCUMENT_AUTO_ARCHIVED,
            MALWARE_UPLOAD_BLOCKED, DLP_PATTERN_DETECTED, DLP_CLASSIFICATION_DOWNGRADE_DENIED,
            DLP_CLASSIFICATION_OVERRIDE_APPROVED,
            DOCUMENT_ACCESS_IMPACT_SIMULATED, SECURITY_RECOMMENDATIONS_VIEWED,
            SECURITY_RECOMMENDATION_STATUS_UPDATED,
            AI_GUARDRAILS_EVALUATED,
            AUDIT_EVENT_CREATED, AUDIT_CHAIN_VERIFIED,
            REQUEST_RECEIVED, REQUEST_OK, REQUEST_DENIED
        resourceType: { type: string }
        resourceId: { type: string, nullable: true }
        result: { $ref: '#/components/schemas/AuditEventResult' }
        reason: { type: string, nullable: true }
        ip: { type: string, nullable: true }
        traceId: { type: string, nullable: true }
        metadata: { type: object, nullable: true, description: 'Structured payload describing the event details (e.g. old/new values, comment content, change diff)' }
        prevHash: { type: string, nullable: true }
        hash: { type: string }
...
    VerifyChainResult:
      type: object
      properties:
        valid: { type: boolean }
        checked: { type: integer }
        firstBrokenIndex: { type: integer, nullable: true }
        firstBrokenEventId: { type: string, nullable: true }
        lastTrustedHash: { type: string, nullable: true }
        message: { type: string, nullable: true }
        epochId: { type: string, nullable: true }
        activeEpoch:
          type: object
          nullable: true
          properties:
            epochId: { type: string }
            status: { type: string }
            valid: { type: boolean }
            checked: { type: integer }
            startedAt: { type: string, format: date-time, nullable: true }
            genesisReason: { type: string, nullable: true }
            previousEpochId: { type: string, nullable: true }
            incidentId: { type: string, nullable: true }
        historicalCompromisedCount: { type: integer, nullable: true }
        compromisedEpochs:
          type: array
          items:
            type: object
            properties:
              epochId: { type: string }
              status: { type: string, enum: [COMPROMISED] }
              startedAt: { type: string, format: date-time, nullable: true }
              endedAt: { type: string, format: date-time, nullable: true }
              incidentId: { type: string, nullable: true }
              firstBrokenIndex: { type: integer, nullable: true }
              firstBrokenEventId: { type: string, nullable: true }
              lastTrustedHash: { type: string, nullable: true }
              reason: { type: string, nullable: true }
    SealAuditChainRequest:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
          minLength: 12
          maxLength: 500
    RiskyDocumentSummary:
      type: object
      properties:
        documentId: { type: string }
        classification: { $ref: '#/components/schemas/ClassificationLevel' }
        accessCount: { type: integer }
        actorCount: { type: integer }
        latestAccessAt: { type: string, format: date-time }
        riskScore: { type: integer, minimum: 0, maximum: 100 }
        reasons:
          type: array
          items: { type: string }
    BehaviorSignalSummary:
      type: object
      properties:
        signalId: { type: string }
        type:
          type: string
          enum: [MASS_CONTENT_ACCESS, DENY_BURST, DESTRUCTIVE_ACTIVITY]
        severity:
          type: string
          enum: [critical, warning, watch]
        actorId: { type: string }
        actionCount: { type: integer }
        documentCount: { type: integer }
        windowStartedAt: { type: string, format: date-time }
        windowEndedAt: { type: string, format: date-time }
        riskScore: { type: integer, minimum: 0, maximum: 100 }
        reasons:
          type: array
          items: { type: string }
    SecurityRecommendationWorkflowStatus:
      type: string
      enum: [OPEN, INVESTIGATING, REVIEWED, RESOLVED]
    SecurityRecommendationWorkflowSummary:
      type: object
      properties:
        status:
          allOf:
            - $ref: '#/components/schemas/SecurityRecommendationWorkflowStatus'
          default: OPEN
        note: { type: string, nullable: true }
        updatedAt: { type: string, format: date-time, nullable: true }
        updatedBy: { type: string, nullable: true }
    SecurityRecommendationWorkflowHistoryEntry:
      type: object
      description: Metadata-only workflow history entry for a security recommendation.
      required: [eventId, status, updatedAt, updatedBy]
      properties:
        eventId: { type: string }
        status: { $ref: '#/components/schemas/SecurityRecommendationWorkflowStatus' }
        note:
          type: string
          description: Optional investigation/review note with content-bearing fields excluded.
        updatedAt: { type: string, format: date-time }
        updatedBy: { type: string }
    UpdateSecurityRecommendationWorkflow:
      type: object
      required: [status]
      properties:
        status: { $ref: '#/components/schemas/SecurityRecommendationWorkflowStatus' }
        note:
          type: string
          maxLength: 500
          description: Optional investigation/review note. Must not contain file content, object keys, presigned URLs, or grant tokens.
    SecurityRecommendationSummary:
      type: object
      description: Deterministic, audit-metadata-only security recommendation.
      properties:
        id: { type: string }
        type:
          type: string
          enum:
            - AUDIT_CHAIN_REVIEW
            - DLP_CLASSIFICATION_REVIEW
            - MALWARE_UPLOAD_REVIEW
            - DOCUMENT_ACCESS_REVIEW
            - ACTOR_ACCESS_REVIEW
        severity:
          type: string
          enum: [critical, warning, info]
        title: { type: string }
        reason: { type: string }
        recommendedAction: { type: string }
        evidence:
          type: array
          items: { type: string }
        affectedDocumentIds:
          type: array
          items: { type: string }
        affectedActorIds:
          type: array
          items: { type: string }
        auditFilters:
          type: object
          description: Audit query filters for opening scoped evidence without exposing file content.
          properties:
            actorId: { type: string }
            action: { type: string }
            actionGroup:
              type: string
              enum: [AUTHORIZED_CONTENT_ACCESS, DESTRUCTIVE_ACTIVITY]
            result: { $ref: '#/components/schemas/AuditEventResult' }
            resourceType: { type: string }
            resourceId: { type: string }
            documentId: { type: string }
            from: { type: string, format: date-time }
            to: { type: string, format: date-time }
        workflow:
          $ref: '#/components/schemas/SecurityRecommendationWorkflowSummary'
    SecuritySummary:
      type: object
      properties:
        chain: { $ref: '#/components/schemas/VerifyChainResult' }
        totals:
          type: object
          properties:
            deniedEvents: { type: integer }
            malwareBlocked: { type: integer }
            dlpDetections: { type: integer }
            downloadDenied: { type: integer }
        repeatedDenyActors:
          type: array
          items:
            type: object
            properties:
              actorId: { type: string }
              denyCount: { type: integer }
        riskyDocuments:
          type: array
          items:
            $ref: '#/components/schemas/RiskyDocumentSummary'
        behaviorSignals:
          type: array
          items:
            $ref: '#/components/schemas/BehaviorSignalSummary'
        recommendations:
          type: array
          items:
            $ref: '#/components/schemas/SecurityRecommendationSummary'
...
paths:
...
  /audit/events:
    post:
      summary: Emit an audit event (internal — called by services)
      operationId: createAuditEvent
      security:
        - bearerAuth: []
        - cookieAuth: []
      tags: [Audit]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [actorId, actorRoles, action, resourceType, result]
              properties:
                eventId: { type: string }
                timestamp: { type: string, format: date-time }
                actorId: { type: string }
                actorRoles: { type: array, items: { type: string } }
                action: { type: string }
                resourceType: { type: string }
                resourceId: { type: string }
                result: { $ref: '#/components/schemas/AuditEventResult' }
                reason: { type: string }
                ip: { type: string }
                traceId: { type: string }
      responses:
        '201':
          description: Event created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditEvent'
        '401':
          description: Unauthorized
...
  /audit/query:
    get:
      summary: Query audit events (compliance officer only)
      operationId: queryAudit
      security:
        - bearerAuth: []
        - cookieAuth: []
      tags: [Audit]
      parameters:
        - in: query
          name: actorId
          schema: { type: string }
        - in: query
          name: action
          schema: { type: string }
        - in: query
          name: actionGroup
          schema:
            type: string
            enum: [AUTHORIZED_CONTENT_ACCESS, DESTRUCTIVE_ACTIVITY]
          description: Expands to a supported evidence action set, such as authorized preview/download grants.
        - in: query
          name: resourceType
          schema: { type: string }
        - in: query
          name: resourceId
          schema: { type: string }
        - in: query
          name: documentId
          schema: { type: string }
          description: Matches direct DOCUMENT events and related events carrying metadata.docId.
        - in: query
          name: aclId
          schema: { type: string }
          description: Matches ACL-level audit events carrying metadata.aclId or metadata.removedAclId.
        - in: query
          name: recommendationId
          schema: { type: string }
          description: Matches SECURITY_RECOMMENDATION workflow events and recommendation metadata references.
        - in: query
          name: result
          schema: { $ref: '#/components/schemas/AuditEventResult' }
        - in: query
          name: from
          schema: { type: string, format: date-time }
        - in: query
          name: to
          schema: { type: string, format: date-time }
        - in: query
          name: limit
          schema: { type: integer, default: 100, maximum: 200 }
      responses:
        '200':
          description: Audit events (sorted by timestamp descending)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AuditEvent'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden — compliance_officer role required
  /audit/verify-chain:
    get:
      summary: Verify integrity of the audit hash chain
      operationId: verifyChain
      security:
        - bearerAuth: []
        - cookieAuth: []
      tags: [Audit]
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 1000
            maximum: 5000
          description: Maximum number of events to check
      responses:
        '200':
          description: Chain verification result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyChainResult'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden — compliance_officer or admin required
  /audit/chain/seal-and-start-epoch:
    post:
      summary: Seal compromised audit epoch and start a new one
      operationId: sealCompromisedAuditChainAndStartEpoch
      description: Preserves the compromised epoch and starts a new active audit chain instead of recomputing historical hashes.
      security:
        - bearerAuth: []
        - cookieAuth: []
      tags: [Audit]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SealAuditChainRequest'
      responses:
        '200':
          description: Recovery incident, previous epoch, new epoch, and audit event
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Chain is currently valid or request is invalid
        '401':
          description: Unauthorized
        '403':
          description: Forbidden — compliance_officer or admin required
  /audit/security-summary:
    get:
      summary: Summarize security audit evidence
      operationId: securitySummary
      security:
        - bearerAuth: []
        - cookieAuth: []
      tags: [Audit]
      responses:
        '200':
          description: Security counters and audit chain status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecuritySummary'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden — compliance_officer or admin required
  /audit/security-recommendations/{id}/workflow:
    patch:
      summary: Update security recommendation workflow state
      operationId: updateSecurityRecommendationWorkflow
      description: |
        Updates the workflow state for a deterministic security recommendation.
        Emits audit action SECURITY_RECOMMENDATION_STATUS_UPDATED with
        resourceType SECURITY_RECOMMENDATION and resourceId equal to the
        recommendation id. Audit metadata stores recommendationId/status/note
        only and must not include file content, object keys, presigned URLs, or
        grant tokens.
      security:
        - bearerAuth: []
        - cookieAuth: []
      tags: [Audit]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
          description: Security recommendation ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSecurityRecommendationWorkflow'
      responses:
        '200':
          description: Workflow update audit event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditEvent'
        '400':
          description: Invalid workflow status or note
        '401':
          description: Unauthorized
        '403':
          description: Forbidden — compliance_officer or admin required
  /audit/security-recommendations/{id}/workflow-history:
    get:
      summary: List security recommendation workflow history
      operationId: getSecurityRecommendationWorkflowHistory
      description: |
        Returns the latest metadata-only workflow events for a deterministic
        security recommendation. Entries expose eventId/status/note/updatedAt/
        updatedBy only and exclude file content, object keys, presigned URLs,
        and grant tokens.
      security:
        - bearerAuth: []
        - cookieAuth: []
      tags: [Audit]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
          description: Security recommendation ID
      responses:
        '200':
          description: Latest workflow history entries, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SecurityRecommendationWorkflowHistoryEntry'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden — compliance_officer or admin required

  # ─── Notifications ─────────────────────────────────────────────────────────
...